+++ crates/client/src/tests.rs
@@ -98,6 +98,7 @@ async fn test_store_and_remove_program() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -142,6 +143,7 @@ async fn test_remove_program_reference_counter() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();