+++ crates/test-cli/src/main.rs
@@ -22,7 +22,7 @@ use entropy_test_cli::run_command;
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     let now = Instant::now();
-    match run_command(None, None, None).await {
+    match run_command(None, None, None, None).await {
         Ok(output) => {
             println!("Success: {}", output.green());
             println!("{}", format!("That took {:?}", now.elapsed()).yellow());