+++ crates/threshold-signature-server/src/user/tests.rs
@@ -95,6 +95,7 @@ use synedrion::{
     k256::ecdsa::{RecoveryId, Signature as k256Signature, VerifyingKey},
     AuxInfo, ThresholdKeyShare,
 };
+use tokio::select;
 use tokio::{
     io::{AsyncRead, AsyncReadExt},
     task::JoinHandle,
@@ -122,7 +123,8 @@ use crate::{
         substrate::{get_oracle_data, query_chain, submit_transaction},
         tests::{
             create_clients, initialize_test_logger, jump_start_network_with_signer, remove_program,
-            run_to_block, setup_client, spawn_testing_validators, unsafe_get,
+            run_to_block, setup_client, spawn_testing_validators, store_program_and_register,
+            unsafe_get, ChainSpecType,
         },
         user::compute_hash,
         validator::get_signer_and_x25519_secret_from_mnemonic,
@@ -169,7 +171,8 @@ async fn test_signature_requests_fail_on_different_conditions() {
     let two = AccountKeyring::Two;
 
     let add_parent_key_to_kvdb = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key_to_kvdb).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key_to_kvdb, ChainSpecType::Integration).await;
 
     // Here we need to use `--chain=integration-tests` and force authoring otherwise we won't be
     // able to get our chain in the right state to be jump started.
@@ -182,37 +185,16 @@ async fn test_signature_requests_fail_on_different_conditions() {
     // for later
     jump_start_network(&entropy_api, &rpc).await;
 
-    // We need to store a program in order to be able to register succesfully
-    let program_hash = test_client::store_program(
-        &entropy_api,
-        &rpc,
-        &two.pair(), // This is our program deployer
-        TEST_PROGRAM_WASM_BYTECODE.to_owned(),
-        vec![],
-        vec![],
-        vec![],
-    )
-    .await
-    .unwrap();
-
-    let (verifying_key, _registered_info) = test_client::register(
-        &entropy_api,
-        &rpc,
-        one.clone().into(), // This is our program modification account
-        subxtAccountId32(two.public().0), // This is our signature request account
-        BoundedVec(vec![ProgramInstance { program_pointer: program_hash, program_config: vec![] }]),
-    )
-    .await
-    .unwrap();
+    // Register the user with a test program
+    let (verifying_key, program_hash) =
+        store_program_and_register(&entropy_api, &rpc, &one.pair(), &two.pair()).await;
 
     // Test: We check that an account with a program succeeds in submiting a signature request
     let (validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
+        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED), verifying_key)
+            .await;
 
     // The account we registered does have a program pointer, so this should succeed
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
-    signature_request.signature_verifying_key = verifying_key.to_vec();
-
     let test_user_res =
         submit_transaction_requests(validator_ips_and_keys.clone(), signature_request.clone(), one)
             .await;
@@ -326,7 +308,8 @@ async fn signature_request_with_derived_account_works() {
     let charlie = AccountKeyring::Charlie;
 
     let add_parent_key_to_kvdb = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key_to_kvdb).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key_to_kvdb, ChainSpecType::Integration).await;
 
     // Here we need to use `--chain=integration-tests` and force authoring otherwise we won't be
     // able to get our chain in the right state to be jump started.
@@ -339,35 +322,13 @@ async fn signature_request_with_derived_account_works() {
     // for later
     jump_start_network(&entropy_api, &rpc).await;
 
-    // We need to store a program in order to be able to register succesfully
-    let program_hash = test_client::store_program(
-        &entropy_api,
-        &rpc,
-        &bob.pair(), // This is our program deployer
-        TEST_PROGRAM_WASM_BYTECODE.to_owned(),
-        vec![],
-        vec![],
-        vec![],
-    )
-    .await
-    .unwrap();
+    // Register the user with a test program
+    let (verifying_key, _program_hash) =
+        store_program_and_register(&entropy_api, &rpc, &charlie.pair(), &bob.pair()).await;
 
-    let (verifying_key, _registered_info) = test_client::register(
-        &entropy_api,
-        &rpc,
-        charlie.clone().into(), // This is our program modification account
-        subxtAccountId32(alice.public().0), // This is our signature request account
-        BoundedVec(vec![ProgramInstance { program_pointer: program_hash, program_config: vec![] }]),
-    )
-    .await
-    .unwrap();
-
-    let (validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
-
-    // We'll use the actual verifying key we registered for the signature request
-    signature_request.signature_verifying_key = verifying_key.to_vec();
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
+    let (validators_info, signature_request, validator_ips_and_keys) =
+        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED), verifying_key)
+            .await;
 
     let signature_request_responses = submit_transaction_requests(
         validator_ips_and_keys.clone(),
@@ -396,7 +357,8 @@ async fn test_signing_fails_if_wrong_participants_are_used() {
     let one = AccountKeyring::Dave;
 
     let add_parent_key = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key, ChainSpecType::Integration).await;
 
     let force_authoring = true;
     let substrate_context = test_node_process_testing_state(force_authoring).await;
@@ -408,8 +370,13 @@ async fn test_signing_fails_if_wrong_participants_are_used() {
 
     let mock_client = reqwest::Client::new();
 
-    let (_validators_info, signature_request, _validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
+    let (_validators_info, signature_request, _validator_ips_and_keys) = get_sign_tx_data(
+        &entropy_api,
+        &rpc,
+        hex::encode(PREIMAGE_SHOULD_SUCCEED),
+        DAVE_VERIFYING_KEY,
+    )
+    .await;
 
     // fails verification tests
     // wrong key for wrong validator
@@ -470,7 +437,8 @@ async fn test_request_limit_are_updated_during_signing() {
     let two = AccountKeyring::Two;
 
     let add_parent_key = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key, ChainSpecType::Integration).await;
 
     let force_authoring = true;
     let substrate_context = test_node_process_testing_state(force_authoring).await;
@@ -480,36 +448,16 @@ async fn test_request_limit_are_updated_during_signing() {
 
     jump_start_network(&entropy_api, &rpc).await;
 
-    let program_hash = test_client::store_program(
-        &entropy_api,
-        &rpc,
-        &two.pair(), // This is our program deployer
-        TEST_PROGRAM_WASM_BYTECODE.to_owned(),
-        vec![],
-        vec![],
-        vec![],
-    )
-    .await
-    .unwrap();
-
-    let (verifying_key, _registered_info) = test_client::register(
-        &entropy_api,
-        &rpc,
-        one.clone().into(), // This is our program modification account
-        subxtAccountId32(two.public().0), // This is our signature request account
-        BoundedVec(vec![ProgramInstance { program_pointer: program_hash, program_config: vec![] }]),
-    )
-    .await
-    .unwrap();
+    // Register the user with a test program
+    let (verifying_key, _program_hash) =
+        store_program_and_register(&entropy_api, &rpc, &one.pair(), &two.pair()).await;
 
     // Test: We check that the rate limiter changes as expected when signature requests are sent
 
     // First we need to get a signature request to populate the KVDB for our verifying key
     let (validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
-
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
-    signature_request.signature_verifying_key = verifying_key.to_vec();
+        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED), verifying_key)
+            .await;
 
     let test_user_res =
         submit_transaction_requests(validator_ips_and_keys.clone(), signature_request.clone(), one)
@@ -592,7 +540,8 @@ async fn test_fails_to_sign_if_non_signing_group_participants_are_used() {
     let two = AccountKeyring::Two;
 
     let add_parent_key = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key, ChainSpecType::Integration).await;
 
     let force_authoring = true;
     let substrate_context = test_node_process_testing_state(force_authoring).await;
@@ -602,30 +551,13 @@ async fn test_fails_to_sign_if_non_signing_group_participants_are_used() {
 
     jump_start_network(&entropy_api, &rpc).await;
 
-    let program_hash = test_client::store_program(
-        &entropy_api,
-        &rpc,
-        &two.pair(), // This is our program deployer
-        TEST_PROGRAM_WASM_BYTECODE.to_owned(),
-        vec![],
-        vec![],
-        vec![],
-    )
-    .await
-    .unwrap();
+    // Register the user with a test program
+    let (verifying_key, _program_hash) =
+        store_program_and_register(&entropy_api, &rpc, &one.pair(), &two.pair()).await;
 
-    let (verifying_key, _registered_info) = test_client::register(
-        &entropy_api,
-        &rpc,
-        one.clone().into(), // This is our program modification account
-        subxtAccountId32(two.public().0), // This is our signature request account
-        BoundedVec(vec![ProgramInstance { program_pointer: program_hash, program_config: vec![] }]),
-    )
-    .await
-    .unwrap();
-
-    let (_validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
+    let (_validators_info, signature_request, validator_ips_and_keys) =
+        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED), verifying_key)
+            .await;
 
     let message_hash = Hasher::keccak(PREIMAGE_SHOULD_SUCCEED);
     let signature_request_account = subxtAccountId32(one.pair().public().0);
@@ -671,12 +603,9 @@ async fn test_fails_to_sign_if_non_signing_group_participants_are_used() {
         encrypted_connection.recv().await.is_err()
     });
 
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
-    signature_request.signature_verifying_key = verifying_key.to_vec();
-
     let test_user_bad_connection_res = submit_transaction_requests(
         vec![validator_ips_and_keys[0].clone()],
-        signature_request.clone(),
+        signature_request,
         one,
     )
     .await;
@@ -684,7 +613,7 @@ async fn test_fails_to_sign_if_non_signing_group_participants_are_used() {
     for res in test_user_bad_connection_res {
         assert_eq!(
             res.unwrap().text().await.unwrap(),
-            "{\"Err\":\"Oneshot timeout error: channel closed\"}"
+            "{\"Err\":\"Subscribe message rejected: NoListener(\\\"no listener\\\")\"}"
         );
     }
 
@@ -703,7 +632,8 @@ async fn test_program_with_config() {
     let two = AccountKeyring::Two;
 
     let add_parent_key = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key, ChainSpecType::Integration).await;
 
     let force_authoring = true;
     let substrate_context = test_node_process_testing_state(force_authoring).await;
@@ -721,6 +651,7 @@ async fn test_program_with_config() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -761,12 +692,8 @@ async fn test_program_with_config() {
     .unwrap();
 
     // Now we'll send off a signature request using the new program
-    let (validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(message)).await;
-
-    // We'll use the actual verifying key we registered for the signature request
-    signature_request.signature_verifying_key = verifying_key.to_vec();
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
+    let (validators_info, signature_request, validator_ips_and_keys) =
+        get_sign_tx_data(&entropy_api, &rpc, hex::encode(message), verifying_key).await;
 
     // Here we check that the signature request was indeed completed successfully
     let signature_request_responses =
@@ -789,45 +716,33 @@ async fn test_jumpstart_network() {
 
     let alice = AccountKeyring::Alice;
 
-    let cxt = test_context_stationary().await;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(false).await;
-    let api = get_api(&cxt.node_proc.ws_url).await.unwrap();
-    let rpc = get_rpc(&cxt.node_proc.ws_url).await.unwrap();
+    let add_parent_key = false;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key, ChainSpecType::Integration).await;
+
+    let force_authoring = true;
+    let substrate_context = test_node_process_testing_state(force_authoring).await;
+
+    let api = get_api(&substrate_context.ws_url).await.unwrap();
+    let rpc = get_rpc(&substrate_context.ws_url).await.unwrap();
 
     let client = reqwest::Client::new();
 
     let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
 
-    let validators_info = vec![
-        entropy_shared::ValidatorInfo {
-            ip_address: b"127.0.0.1:3001".to_vec(),
-            x25519_public_key: X25519_PUBLIC_KEYS[0],
-            tss_account: TSS_ACCOUNTS[0].clone().encode(),
-        },
-        entropy_shared::ValidatorInfo {
-            ip_address: b"127.0.0.1:3002".to_vec(),
-            x25519_public_key: X25519_PUBLIC_KEYS[1],
-            tss_account: TSS_ACCOUNTS[1].clone().encode(),
-        },
-        entropy_shared::ValidatorInfo {
-            ip_address: b"127.0.0.1:3003".to_vec(),
-            x25519_public_key: X25519_PUBLIC_KEYS[2],
-            tss_account: TSS_ACCOUNTS[2].clone().encode(),
-        },
-    ];
-    let onchain_user_request = OcwMessageDkg {
-        sig_request_accounts: vec![NETWORK_PARENT_KEY.encode()],
-        block_number,
-        validators_info,
-    };
-
     put_jumpstart_request_on_chain(&api, &rpc, &alice).await;
 
     run_to_block(&rpc, block_number + 1).await;
 
+    let selected_validators_query = entropy::storage().registry().jumpstart_dkg(block_number);
+    let validators_info =
+        query_chain(&api, &rpc, selected_validators_query, None).await.unwrap().unwrap();
+    let validators_info: Vec<_> = validators_info.into_iter().map(|v| v.0).collect();
+    let onchain_user_request = OcwMessageDkg { block_number, validators_info };
+
     // succeeds
     let response_results = join_all(
-        vec![3002, 3003]
+        vec![3002, 3003, 3004]
             .iter()
             .map(|port| {
                 client
@@ -929,6 +844,7 @@ async fn test_compute_hash() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -1002,7 +918,8 @@ async fn test_fail_infinite_program() {
     let two = AccountKeyring::Two;
 
     let add_parent_key = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key, ChainSpecType::Integration).await;
 
     let force_authoring = true;
     let substrate_context = test_node_process_testing_state(force_authoring).await;
@@ -1020,6 +937,7 @@ async fn test_fail_infinite_program() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -1035,12 +953,8 @@ async fn test_fail_infinite_program() {
     .unwrap();
 
     // Now we'll send off a signature request using the new program
-    let (_validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
-
-    // We'll use the actual verifying key we registered for the signature request
-    signature_request.signature_verifying_key = verifying_key.to_vec();
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
+    let (_validators_info, signature_request, validator_ips_and_keys) =
+        get_sign_tx_data(&api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED), verifying_key).await;
 
     let test_infinite_loop =
         submit_transaction_requests(validator_ips_and_keys.clone(), signature_request.clone(), one)
@@ -1085,7 +999,8 @@ async fn test_device_key_proxy() {
     let two = AccountKeyring::Two;
 
     let add_parent_key_to_kvdb = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key_to_kvdb).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key_to_kvdb, ChainSpecType::Integration).await;
 
     // Here we need to use `--chain=integration-tests` and force authoring otherwise we won't be
     // able to get our chain in the right state to be jump started.
@@ -1107,6 +1022,7 @@ async fn test_device_key_proxy() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -1177,11 +1093,9 @@ async fn test_device_key_proxy() {
 
     // Now we'll send off a signature request using the new program with auxilary data
     let (validators_info, mut signature_request, validator_ips_and_keys) =
-        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED)).await;
+        get_sign_tx_data(&entropy_api, &rpc, hex::encode(PREIMAGE_SHOULD_SUCCEED), verifying_key)
+            .await;
 
-    // We'll use the actual verifying key we registered for the signature request
-    signature_request.signature_verifying_key = verifying_key.to_vec();
-    signature_request.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
     signature_request.auxilary_data = auxilary_data;
 
     let test_user_res =
@@ -1223,7 +1137,8 @@ async fn test_faucet() {
     let two = AccountKeyring::Eve;
     let alice = AccountKeyring::Alice;
 
-    let (validator_ips, _validator_ids) = spawn_testing_validators(false).await;
+    let (validator_ips, _validator_ids) =
+        spawn_testing_validators(false, ChainSpecType::Development).await;
     let substrate_context = test_node_process_testing_state(true).await;
     let entropy_api = get_api(&substrate_context.ws_url).await.unwrap();
     let rpc = get_rpc(&substrate_context.ws_url).await.unwrap();
@@ -1260,6 +1175,7 @@ async fn test_faucet() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -1386,7 +1302,7 @@ async fn test_faucet() {
 
 #[tokio::test]
 #[serial]
-async fn test_new_registration_flow() {
+async fn test_registration_flow() {
     initialize_test_logger().await;
     clean_tests();
 
@@ -1395,7 +1311,8 @@ async fn test_new_registration_flow() {
     let charlie = AccountKeyring::Charlie;
 
     let add_parent_key_to_kvdb = true;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key_to_kvdb).await;
+    let (_validator_ips, _validator_ids) =
+        spawn_testing_validators(add_parent_key_to_kvdb, ChainSpecType::Integration).await;
 
     // Here we need to use `--chain=integration-tests` and force authoring otherwise we won't be
     // able to get our chain in the right state to be jump started.
@@ -1423,6 +1340,7 @@ async fn test_new_registration_flow() {
         vec![],
         vec![],
         vec![],
+        0u8,
     )
     .await
     .unwrap();
@@ -1583,6 +1501,7 @@ pub async fn get_sign_tx_data(
     api: &OnlineClient<EntropyConfig>,
     rpc: &LegacyRpcMethods<EntropyConfig>,
     message: String,
+    signature_verifying_key: [u8; 33],
 ) -> (Vec<ValidatorInfo>, UserSignatureRequest, Vec<(String, [u8; 32])>) {
     let validators_info = get_signers_from_chain(api, rpc).await.unwrap();
 
@@ -1593,9 +1512,9 @@ pub async fn get_sign_tx_data(
             Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         ]),
         validators_info: validators_info.clone(),
-        block_number: 0,
+        block_number: rpc.chain_get_header(None).await.unwrap().unwrap().number,
         hash: HashingAlgorithm::Keccak,
-        signature_verifying_key: DAVE_VERIFYING_KEY.to_vec(),
+        signature_verifying_key: signature_verifying_key.to_vec(),
     };
 
     let validator_ips_and_keys =