+++ pallets/programs/src/tests.rs
@@ -30,12 +30,14 @@ fn set_program() {
         let configuration_schema = vec![14u8];
         let auxiliary_data_schema = vec![15u8];
         let oracle_data_pointer = vec![16u8];
+        let version_number = 0u8;
         let too_long = vec![1u8, 2u8, 3u8, 4u8, 5u8];
         let mut hash_input: Vec<u8> = vec![];
         hash_input.extend(&program);
         hash_input.extend(&configuration_schema);
         hash_input.extend(&auxiliary_data_schema);
         hash_input.extend(&oracle_data_pointer);
+        hash_input.extend(&vec![version_number]);
 
         let program_hash = <Test as frame_system::Config>::Hashing::hash(&hash_input);
         // can't pay deposit
@@ -45,7 +47,8 @@ fn set_program() {
                 program.clone(),
                 configuration_schema.clone(),
                 auxiliary_data_schema.clone(),
-                oracle_data_pointer.clone()
+                oracle_data_pointer.clone(),
+                version_number
             ),
             BalancesError::<Test>::InsufficientBalance
         );
@@ -57,7 +60,8 @@ fn set_program() {
             program.clone(),
             configuration_schema.clone(),
             auxiliary_data_schema.clone(),
-            oracle_data_pointer.clone()
+            oracle_data_pointer.clone(),
+            version_number
         ));
         let program_result = ProgramInfo {
             bytecode: program.clone(),
@@ -66,6 +70,7 @@ fn set_program() {
             oracle_data_pointer: oracle_data_pointer.clone(),
             deployer: PROGRAM_MODIFICATION_ACCOUNT,
             ref_counter: 0u128,
+            version_number,
         };
         assert_eq!(
             ProgramsPallet::programs(program_hash).unwrap(),
@@ -88,6 +93,7 @@ fn set_program() {
                 configuration_schema.clone(),
                 auxiliary_data_schema.clone(),
                 oracle_data_pointer.clone(),
+                version_number
             ),
             Error::<Test>::ProgramAlreadySet
         );
@@ -100,6 +106,7 @@ fn set_program() {
                 configuration_schema.clone(),
                 auxiliary_data_schema.clone(),
                 oracle_data_pointer.clone(),
+                version_number
             ),
             Error::<Test>::TooManyProgramsOwned
         );
@@ -111,6 +118,7 @@ fn set_program() {
                 configuration_schema,
                 auxiliary_data_schema.clone(),
                 oracle_data_pointer.clone(),
+                version_number
             ),
             Error::<Test>::ProgramLengthExceeded
         );
@@ -124,11 +132,13 @@ fn remove_program() {
         let configuration_schema = vec![14u8];
         let auxiliary_data_schema = vec![15u8];
         let oracle_data_pointer = vec![16u8];
+        let version_number = 0u8;
         let mut hash_input: Vec<u8> = vec![];
         hash_input.extend(&program);
         hash_input.extend(&configuration_schema);
         hash_input.extend(&auxiliary_data_schema);
         hash_input.extend(&oracle_data_pointer);
+        hash_input.extend(&vec![version_number]);
         let program_hash = <Test as frame_system::Config>::Hashing::hash(&hash_input);
 
         // no program
@@ -147,7 +157,8 @@ fn remove_program() {
             program.clone(),
             configuration_schema.clone(),
             auxiliary_data_schema.clone(),
-            oracle_data_pointer.clone()
+            oracle_data_pointer.clone(),
+            version_number
         ));
         assert_eq!(
             ProgramsPallet::owned_programs(PROGRAM_MODIFICATION_ACCOUNT),
@@ -196,6 +207,7 @@ fn remove_program_fails_ref_count() {
         let configuration_schema = vec![14u8];
         let auxiliary_data_schema = vec![15u8];
         let oracle_data_pointer = vec![16u8];
+        let version_number = 0u8;
 
         Programs::<Test>::insert(
             program_hash,
@@ -206,6 +218,7 @@ fn remove_program_fails_ref_count() {
                 oracle_data_pointer,
                 deployer: PROGRAM_MODIFICATION_ACCOUNT,
                 ref_counter: 1u128,
+                version_number,
             },
         );