+++ pallets/propagation/src/tests.rs
@@ -33,13 +33,7 @@ fn knows_how_to_mock_several_http_calls() {
             uri: "http://localhost:3001/generate_network_key".into(),
             sent: true,
             response: Some([].to_vec()),
-            body: [
-                0, 0, 0, 0, 0, 8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 10, 32, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4,
-                11, 32, 4, 0, 0, 0, 0, 0, 0, 0,
-            ]
-            .to_vec(),
+            body: [0, 0, 0, 0, 0].to_vec(),
             ..Default::default()
         });